+++ openapi_schema_pydantic/v3/v3_0_3/contact.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from pydantic import BaseModel, AnyUrl
+from pydantic import AnyUrl, BaseModel, Extra
 
 
 class Contact(BaseModel):
@@ -26,6 +26,7 @@ class Contact(BaseModel):
     """
 
     class Config:
+        extra = Extra.allow
         schema_extra = {
             "examples": [
                 {"name": "API Support", "url": "http://www.example.com/support", "email": "support@example.com"}