+++ openapi_schema_pydantic/v3/v3_0_3/schema.py
@@ -471,8 +471,8 @@ class Schema(BaseModel):
     """
 
     class Config:
-        allow_population_by_field_name = True
         extra = Extra.allow
+        allow_population_by_field_name = True
         schema_extra = {
             "examples": [
                 {"type": "string", "format": "email"},