+++ openapi_schema_pydantic/v3/v3_0_3/server_variable.py
@@ -1,6 +1,6 @@
 from typing import List, Optional
 
-from pydantic import BaseModel
+from pydantic import BaseModel, Extra
 
 
 class ServerVariable(BaseModel):
@@ -26,3 +26,6 @@ class ServerVariable(BaseModel):
     An optional description for the server variable.
     [CommonMark syntax](https://spec.commonmark.org/) MAY be used for rich text representation.
     """
+
+    class Config:
+        extra = Extra.allow