+++ openapi_schema_pydantic/v3/v3_0_3/tag.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from pydantic import BaseModel
+from pydantic import BaseModel, Extra
 
 from .external_documentation import ExternalDocumentation
 
@@ -28,4 +28,5 @@ class Tag(BaseModel):
     """
 
     class Config:
+        extra = Extra.allow
         schema_extra = {"examples": [{"name": "pet", "description": "Pets operations"}]}